update author for 1 Samuel and 2 Samuel
The book of 1 Samuel was written by Samuel, Gad and Nathan as explained in 1 Chronicles 29:29. It'd be easy to assume that 2 Samuel was also written by all 3 authors but 1 Samuel 25:1 speaks to Samuel's death so it's safer to assume that only Gad and Nathan participated in the writing of 2 Samuel.